+++ src/components/Dashboard/Settings/Settings.vue
@@ -1,14 +1,35 @@
 <script setup>
 import { record } from '../../../store/record';
 import { saveAs } from 'file-saver';
-import { DocumentArrowDownIcon } from '@heroicons/vue/24/outline';
+import { DocumentArrowDownIcon, LockClosedIcon } from '@heroicons/vue/24/outline';
+import { encrypt } from '../../../helpers/encrypto';
+import PassphraseModal from './PassphraseModal.vue';
+import { ref } from 'vue';
 
-const downloadHealthRecordFile = () => {
+const passphraseModalOpen = ref(false);
+
+const downloadHealthRecordFile = (data) => {
   const fileName  = 'healthRecord.json';
-  const file      = new Blob([JSON.stringify(record.value)], {
+  const file      = new Blob([JSON.stringify(data)], {
     type: 'application/json'
   });
   saveAs(file, fileName);
+}
+
+const passphraseSubmitted = async (passphrase) => {
+  passphraseModalOpen.value = false;
+  const encryptedData = await encrypt(JSON.stringify(record.value), passphrase);
+
+  if (! encryptedData) {
+    return alert('Could not encrypt data.');
+  }
+
+  const data = {
+    encrypted: true,
+    data: encryptedData
+  }
+
+  downloadHealthRecordFile(data);
 }
 
 const appVersion = APP_VERSION;
@@ -29,8 +50,15 @@ const appVersion = APP_VERSION;
       </section>
       <section class="grid grid-flow-col grid-cols-[auto_min-content]  items-center md:mx-0 border-b p-3 px-4 md:px-6">
         <h3 class="font-semibold">Data file</h3>
-        <button class="btn" @click="downloadHealthRecordFile">
+        <button class="btn" @click="downloadHealthRecordFile(record)">
           <DocumentArrowDownIcon />
+          Download
+        </button>
+      </section>
+      <section class="grid grid-flow-col grid-cols-[auto_min-content]  items-center md:mx-0 border-b p-3 px-4 md:px-6">
+        <h3 class="font-semibold">Encrypted data file</h3>
+        <button class="btn" @click="passphraseModalOpen = true">
+          <LockClosedIcon />
           Download
         </button>
       </section>
@@ -39,6 +67,7 @@ const appVersion = APP_VERSION;
         <span>{{ appVersion }}</span>
       </section>
     </div>
+    <PassphraseModal v-if="passphraseModalOpen" @close="passphraseModalOpen = false" @passphraseSubmitted="passphraseSubmitted" />
     <p class="text-center self-end text-xs text-gray-300 mt-6 mb-3">HealthRecord Copyright &copy; {{ new Date().getFullYear() }}</p>
   </div>
 </template>