+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.3] - 2023-09-18
+
+### Added
+- User can download passphrase encrypted data file
+- User can load passphrase encrypted data file
+
+### Changed
+- UUID generation for objects (using browser built-in crypto.randomUUID)
+
+### Fixed
+
+
 ## [0.1.2] - 2023-09-16
 
 ### Added
@@ -20,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Vital chart on Person page more reactive with computed data
+
 
 ## [0.1.1] - 2023-09-15